[#1942] Factor out wire format in testing backend
